requirements.txt
pdfplumber
python-docx
pillow
pytesseract
spacy
regex
scikit-learn
joblib
